# ai-specs/specs/api-spec.yml
openapi: 3.0.0
info:
  title: DocFlow API
  description: |
    REST API for DocFlow — Intelligent Document Management System (DMS).
    
    DocFlow is a modular, API-First document management infrastructure designed to serve as a central repository
    for advanced document lifecycle management, versioning, and semantic search capabilities.
    
    The platform implements a **microservices architecture** with the following core domains:
    - **Identity & Authentication**: Multi-organization support with JWT-based stateless authentication (Identity Service)
    - **Document Management**: Core document lifecycle management, versioning, and storage (Document-Core Service)
    - **API Gateway**: Single entry point routing to backend microservices (Gateway Service)
    - **Audit & Compliance**: Activity tracking and audit logging (AuditLog Service)
    
    ## Architecture & Design Principles
    - **Hexagonal Architecture** (Ports & Adapters): Clear separation between domain logic, application services, and infrastructure
    - **Domain-Driven Design (DDD)**: Business-centric modeling with ubiquitous language
    - **RBAC (Role-Based Access Control)**: Fine-grained permission management across resources
    - **Stateless Authentication**: JWT tokens with organization context
    
    ## Tech Stack
    - **Backend**: Java 21, Spring Boot 3.4.2+, Spring Cloud Gateway, Spring WebFlux
    - **Frontend**: React 19+, Vite, TypeScript, Tailwind CSS, Zustand (state management)
...
    
...
servers:
...
paths:
...
  /api/carpetas/{carpetaId}/mi-permiso:
    get:
      summary: Get effective folder permission
      description: Returns the effective permission for the authenticated user (direct or inherited).
      tags:
        - ACL - Folder Permissions
      security:
        - bearerAuth: []
      parameters:
        - name: carpetaId
          in: path
          required: true
          schema:
            type: integer
          description: Folder ID
      responses:
        '200':
          description: Effective permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermisoEfectivoResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Folder not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    PermisoEfectivoResponse:
      type: object
      properties:
        nivel_acceso:
          type: string
          enum: [LECTURA, ESCRITURA, ADMINISTRACION]
          description: Effective access level
        es_heredado:
          type: boolean
          description: Indicates if the permission is inherited
        carpeta_origen_id:
          type: integer
          description: Folder ID where the permission originates
        carpeta_origen_nombre:
          type: string
          description: Folder name where the permission originates
        ruta_herencia:
          type: array
          items:
            type: string
          nullable: true
          description: Inheritance path from origin to target folder
      required:
        - nivel_acceso
        - es_heredado
        - carpeta_origen_id
        - carpeta_origen_nombre
    
    # Document Permission Schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code/identifier
        message:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        path:
          type: string
          description: Request path that caused the error
        status:
          type: integer
          description: HTTP status code
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: Additional error details (validation errors, etc.)
      required:
        - error
        - message
        - status
...
security:
  - bearerAuth: []
...